iio: ad9144: Apply SERDES optimization values for AD9152
Similar to the AD9144 the AD9152 datasheet lists a set of SERDES
optimization values that should be written to the device for optimum
performance.

Most of the optimization values are identical to the AD9144 optimization
values. But a few of them differ and depend on the VCO frequency. Write
those that are identical out unconditionally.

Also use regmap_multi_reg_write() instead of writing each register
individually.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>