iio: ad9144: Add infrastructure for PLL support
Add all the necessary code to configure the internal PLL inside the AD9144
and use it to generate the sample rate.

For the time being this feature is not exposed yet though since there is
no good interface for it. But if somebody is inclined to modify the drivers
probe function it is very easy to enable by setting the pll_frequency and
pll_enable fields of the ad9144 driver state struct.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>