iio: ad9144: Add support for updating SYSREF frequency
The frequency of the SYSREF signal needs to be a integer division of the
LMFC frequency of the converter. The LMFC depends on the sample rate as
well as the JESD204 link settings.

At the moment it is up to somebody with system knowledge to configure the
SYSREF frequency beforehand so it matches the requirements.

Add some code to the driver that verifies that the currently selected
SYSREF frequency is OK and if it is not reconfigure the clock rate of the
SYSREF provider.

Note that this will not guarantee deterministic latency as reprogramming
SYSREF clock might introduce an arbitrary phase shift to the DAC clock.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>